+++ src/pages/auth/ForgotPassword.tsx
@@ -7,6 +7,7 @@ import { RHFSubmitButton } from "../../components/hook-form/RHFSubmitButton";
 import { useRouter } from "../../hooks/useRouter";
 import { paths } from "../../routes/paths";
 import { AuthHeader } from "../../sections/auth/AuthHeader";
+import { AuthSocial } from "../../sections/auth/AuthSocial";
 
 export type FormValues = {
   username: string;
@@ -19,30 +20,6 @@ const schema = yup.object().shape({
 const defaultValues: FormValues = {
   username: "temp@gmail.com",
 };
-
-type LoginOption = {
-  id: number;
-  name: string;
-  icon: string;
-};
-
-const loginOptions: LoginOption[] = [
-  {
-    id: 1,
-    name: "Facebook",
-    icon: "/icons/ic_facebook.svg",
-  },
-  {
-    id: 2,
-    name: "Google",
-    icon: "/icons/ic_google.svg",
-  },
-  {
-    id: 3,
-    name: "Apple",
-    icon: "/icons/ic_apple.svg",
-  },
-];
 
 export const ForgotPassword: React.FC = () => {
   const router = useRouter();
@@ -57,9 +34,7 @@ export const ForgotPassword: React.FC = () => {
   const onSubmit = (data: FormValues) => {
     console.log(data);
 
-    handlePush(
-      `/${paths.auth.root}/${paths.auth.verificationPassword}`
-    );
+    handlePush(`/${paths.auth.root}/${paths.auth.verificationPassword}`);
   };
 
   const handlePush = (path: string) => {
@@ -88,29 +63,7 @@ export const ForgotPassword: React.FC = () => {
               <RHFSubmitButton color="black" />
             </div>
 
-            <div className="flex items-center justify-between">
-              <div className="w-full h-0.125 bg-gray-500" />
-              <p className="text-base text-gray-500 font-semibold text-nowrap mx-4">
-                Veya Şununla Giriş Yapın
-              </p>
-              <div className="w-full h-0.125 bg-gray-500" />
-              <div />
-            </div>
-
-            <div className="flex items-center gap-4 w-full">
-              {loginOptions.map((option) => (
-                <div
-                  key={`${option.id}${option.name}`}
-                  className="flex items-center justify-center py-4 w-full cursor-pointer border border-gray-200 rounded-lg "
-                >
-                  <img
-                    className="w-6 h-6 select-none"
-                    src={option.icon}
-                    alt={option.name}
-                  />
-                </div>
-              ))}
-            </div>
+            <AuthSocial />
 
             {/* <RHFFormValues /> */}
           </form>
