+++ src/pages/auth/Login.tsx
@@ -11,6 +11,7 @@ import { useRouter } from "../../hooks/useRouter";
 import { paths } from "../../routes/paths";
 import { AuthHeader } from "../../sections/auth/AuthHeader";
 import { RHFCheckBox } from "../../sections/personal-inforations/rhf-components/RHFCheckbox";
+import { AuthSocial } from "../../sections/auth/AuthSocial";
 
 export type FormValues = {
   username: string;
@@ -32,30 +33,6 @@ const defaultValues: FormValues = {
   password: "83r5^_",
   isCheck: false,
 };
-
-type LoginOption = {
-  id: number;
-  name: string;
-  icon: string;
-};
-
-const loginOptions: LoginOption[] = [
-  {
-    id: 1,
-    name: "Facebook",
-    icon: "/icons/ic_facebook.svg",
-  },
-  {
-    id: 2,
-    name: "Google",
-    icon: "/icons/ic_google.svg",
-  },
-  {
-    id: 3,
-    name: "Apple",
-    icon: "/icons/ic_apple.svg",
-  },
-];
 
 export const Login: React.FC = () => {
   const router = useRouter();
@@ -122,29 +99,7 @@ export const Login: React.FC = () => {
               </p>
             </div>
 
-            <div className="flex items-center justify-between">
-              <div className="w-full h-0.125 bg-gray-500" />
-              <p className="text-base text-gray-500 font-semibold text-nowrap mx-4">
-                Veya Şununla Giriş Yapın
-              </p>
-              <div className="w-full h-0.125 bg-gray-500" />
-              <div />
-            </div>
-
-            <div className="flex items-center gap-4 w-full">
-              {loginOptions.map((option) => (
-                <div
-                  key={`${option.id}${option.name}`}
-                  className="flex items-center justify-center py-4 w-full cursor-pointer border border-gray-200 rounded-lg "
-                >
-                  <img
-                    className="w-6 h-6 select-none"
-                    src={option.icon}
-                    alt={option.name}
-                  />
-                </div>
-              ))}
-            </div>
+            <AuthSocial />
 
             {/* <RHFFormValues /> */}
           </form>
